+++ Classes/Controller/CrawlerController.php
@@ -31,6 +31,7 @@ use AOE\Crawler\Service\ConfigurationService;
 use AOE\Crawler\Service\PageService;
 use AOE\Crawler\Service\UrlService;
 use AOE\Crawler\Utility\SignalSlotUtility;
+use AOE\Crawler\Value\QueueRow;
 use PDO;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
@@ -349,6 +350,7 @@ class CrawlerController implements LoggerAwareInterface
             $duplicateTrack[$uKey] = true;
         }
 
+        // Todo: Find a better option to have this correct in both backend (<br>) and cli (<new line>)
         return implode('<br>', $urlLog);
     }
 
@@ -766,7 +768,7 @@ class CrawlerController implements LoggerAwareInterface
      * @param boolean $downloadCrawlUrls If set (and submitcrawlUrls is false) will fill $downloadUrls with entries)
      * @param array $incomingProcInstructions Array of processing instructions
      * @param array $configurationSelection Array of configuration keys
-     * @return string
+     * @return array
      */
     public function getPageTreeAndUrls(
         $id,
@@ -808,9 +810,9 @@ class CrawlerController implements LoggerAwareInterface
             $tree->getTree($id, $depth, '');
         }
 
-        // Traverse page tree:
-        $code = '';
+        $queueRows = [];
 
+        // Traverse page tree:
         foreach ($tree->tree as $data) {
             $this->MP = false;
 
@@ -826,10 +828,11 @@ class CrawlerController implements LoggerAwareInterface
                 $mountTree->getTree($mountpage[0]['mount_pid'], $depth);
 
                 foreach ($mountTree->tree as $mountData) {
-                    $code .= $this->drawURLs_addRowsForPage(
+                    $queueRows = array_merge($queueRows, $this->drawURLs_addRowsForPage(
                         $mountData['row'],
-                        $mountData['HTML'] . BackendUtility::getRecordTitle('pages', $mountData['row'], true)
-                    );
+                        BackendUtility::getRecordTitle('pages', $mountData['row'], true),
+                        (string) $data['HTML']
+                    ));
                 }
 
                 // replace page when mount_pid_ol is enabled
@@ -841,20 +844,21 @@ class CrawlerController implements LoggerAwareInterface
                 }
             }
 
-            $code .= $this->drawURLs_addRowsForPage(
+            $queueRows = array_merge($queueRows, $this->drawURLs_addRowsForPage(
                 $data['row'],
-                $data['HTML'] . BackendUtility::getRecordTitle('pages', $data['row'], true)
-            );
+                BackendUtility::getRecordTitle('pages', $data['row'], true),
+                (string) $data['HTML']
+            ));
         }
 
-        return $code;
+        return $queueRows;
     }
 
     /**
      * Create the rows for display of the page tree
      * For each page a number of rows are shown displaying GET variable configuration
      */
-    public function drawURLs_addRowsForPage(array $pageRow, string $pageTitle): string
+    public function drawURLs_addRowsForPage(array $pageRow, string $pageTitle, string $pageTitleHTML = ''): array
     {
         $skipMessage = '';
 
@@ -864,15 +868,19 @@ class CrawlerController implements LoggerAwareInterface
 
         // Traverse parameter combinations:
         $c = 0;
-        $content = '';
+
+        $queueRowCollection = [];
+
         if (! empty($configurations)) {
             foreach ($configurations as $confKey => $confArray) {
 
                 // Title column:
                 if (! $c) {
-                    $titleClm = '<td rowspan="' . count($configurations) . '">' . $pageTitle . '</td>';
+                    $queueRow = new QueueRow($pageTitle);
+                    $queueRow->setPageTitleHTML($pageTitleHTML);
                 } else {
-                    $titleClm = '';
+                    $queueRow = new QueueRow();
+                    $queueRow->setPageTitleHTML($pageTitleHTML);
                 }
 
                 if (! in_array($pageRow['uid'], $this->configurationService->expandExcludeString($confArray['subCfg']['exclude'] ?? ''), true)) {
@@ -911,47 +919,43 @@ class CrawlerController implements LoggerAwareInterface
                     $paramExpanded .= 'Comb: ' . implode('*', $calcAccu) . '=' . $calcRes;
 
                     // Options
-                    $optionValues = '';
-                    if ($confArray['subCfg']['userGroups']) {
-                        $optionValues .= 'User Groups: ' . $confArray['subCfg']['userGroups'] . '<br/>';
+                    $queueRowOptionCollection = [];
+                    if ($confArray['subCfg']['userGroups'] ?? false) {
+                        $queueRowOptionCollection[] = 'User Groups: ' . $confArray['subCfg']['userGroups'];
                     }
-                    if ($confArray['subCfg']['procInstrFilter']) {
-                        $optionValues .= 'ProcInstr: ' . $confArray['subCfg']['procInstrFilter'] . '<br/>';
+                    if ($confArray['subCfg']['procInstrFilter'] ?? false) {
+                        $queueRowOptionCollection[] = 'ProcInstr: ' . $confArray['subCfg']['procInstrFilter'];
                     }
 
-                    // Compile row:
-                    $content .= '
-                        <tr>
-                            ' . $titleClm . '
-                            <td>' . htmlspecialchars($confKey) . '</td>
-                            <td>' . nl2br(htmlspecialchars(rawurldecode(trim(str_replace('&', chr(10) . '&', GeneralUtility::implodeArrayForUrl('', $confArray['paramParsed'])))))) . '</td>
-                            <td>' . $paramExpanded . '</td>
-                            <td nowrap="nowrap">' . $urlList . '</td>
-                            <td nowrap="nowrap">' . $optionValues . '</td>
-                            <td nowrap="nowrap">' . DebugUtility::viewArray($confArray['subCfg']['procInstrParams.']) . '</td>
-                        </tr>';
+                    // Remove empty array entries;
+                    $queueRowOptionCollection = array_filter($queueRowOptionCollection);
+
+                    $parameterConfig = nl2br(htmlspecialchars(rawurldecode(trim(str_replace('&', chr(10) . '&', GeneralUtility::implodeArrayForUrl('', $confArray['paramParsed'] ?? []))))));
+                    $queueRow->setValuesExpanded($paramExpanded);
+                    $queueRow->setConfigurationKey($confKey);
+                    $queueRow->setUrls($urlList);
+                    $queueRow->setOptions($queueRowOptionCollection);
+                    $queueRow->setParameters(DebugUtility::viewArray($confArray['subCfg']['procInstrParams.'] ?? []));
+                    $queueRow->setParameterConfig($parameterConfig);
+
+                    $queueRowCollection[] = $queueRow;
                 } else {
-                    $content .= '<tr>
-                            ' . $titleClm . '
-                            <td>' . htmlspecialchars($confKey) . '</td>
-                            <td colspan="5"><em>No entries</em> (Page is excluded in this configuration)</td>
-                        </tr>';
+                    $queueRow->setConfigurationKey($confKey);
+                    $queueRow->setMessage('(Page is excluded in this configuration)');
+                    $queueRowCollection[] = $queueRow;
                 }
 
                 $c++;
             }
         } else {
             $message = ! empty($skipMessage) ? ' (' . $skipMessage . ')' : '';
-
-            // Compile row:
-            $content .= '
-                <tr>
-                    <td>' . $pageTitle . '</td>
-                    <td colspan="6"><em>No entries</em>' . $message . '</td>
-                </tr>';
+            $queueRow = new QueueRow($pageTitle);
+            $queueRow->setPageTitleHTML($pageTitleHTML);
+            $queueRow->setMessage($message);
+            $queueRowCollection[] = $queueRow;
         }
 
-        return $content;
+        return $queueRowCollection;
     }
 
     /**